+++ src/pquant/core/train.py
@@ -1,12 +1,6 @@
-import keras
+from pquant.core.utils import get_backend
 
 
 def iterative_train(model, config, train_func, valid_func, **kwargs):
-    if keras.backend.backend() == "torch":
-        from pquant.core.torch_impl.train_torch import iterative_train_torch
-
-        return iterative_train_torch(model, config, train_func, valid_func, **kwargs)
-    else:
-        from pquant.core.tf_impl.train_tf import iterative_train_tf
-
-        return iterative_train_tf(model, config, train_func, valid_func, **kwargs)
+    backend = get_backend()
+    return backend.iterative_train(model, config, train_func, valid_func, **kwargs)
