+++ src/pquant/core/utils.py
@@ -1,7 +1,9 @@
 import os
 
+import keras
 import yaml
 
+from pquant.core.backend_interface import BackendInterface
 from pquant.pruning_methods.activation_pruning import ActivationPruning
 from pquant.pruning_methods.autosparse import AutoSparse
 from pquant.pruning_methods.cs import ContinuousSparsification
@@ -9,6 +11,16 @@ from pquant.pruning_methods.dst import DST
 from pquant.pruning_methods.pdp import PDP
 from pquant.pruning_methods.wanda import Wanda
 from pquant.pruning_methods.mdmm import MDMM
+
+
+def get_backend() -> BackendInterface:
+    from pquant.core.tf_backend import TFBackend
+    from pquant.core.torch_backend import TorchBackend
+
+    if keras.backend.backend() == "torch":
+        return TorchBackend()
+    else:
+        return TFBackend()
 
 
 def get_pruning_layer(config, layer_type):
